requirements.txt
Flask==2.3.3
PyYAML>=6.0